+++ Products/TinyMCE/browser/atanchors.py
@@ -5,6 +5,7 @@ from zope.interface import implements
 
 try:
     from lxml.html import fromstring
+    fromstring     # pyflakes
     SEARCHPATTERN = "a"
 except ImportError:
     from elementtree import HTMLTreeBuilder